+++ lib-e131/src/node/e131bridgeprint.cpp
@@ -56,7 +56,7 @@ void E131Bridge::Print() {
 		for (uint32_t nPortIndex = 0; nPortIndex < e131bridge::MAX_PORTS; nPortIndex++) {
 			uint16_t nUniverse;
 			if (GetUniverse(nPortIndex, nUniverse, lightset::PortDir::OUTPUT)) {
-				printf("  Port %-2d %-4u %s\n", nPortIndex, nUniverse, lightset::get_merge_mode(m_OutputPort[nPortIndex].mergeMode, true));
+				printf("  Port %-2u %-4u %s\n", static_cast<unsigned int>(nPortIndex), static_cast<unsigned int>(nUniverse), lightset::get_merge_mode(m_OutputPort[nPortIndex].mergeMode, true));
 			}
 		}
 	}
@@ -68,7 +68,7 @@ void E131Bridge::Print() {
 		for (uint32_t nPortIndex = 0; nPortIndex < e131bridge::MAX_PORTS; nPortIndex++) {
 			uint16_t nUniverse;
 			if (GetUniverse(nPortIndex, nUniverse, lightset::PortDir::INPUT)) {
-				printf("  Port %-2d %-4u %-3u\n", nPortIndex, nUniverse, GetPriority(nPortIndex));
+				printf("  Port %-2u %-4u %-3u\n", static_cast<unsigned int>(nPortIndex), static_cast<unsigned int>(nUniverse), GetPriority(nPortIndex));
 			}
 		}
 	}
